experimental implementation of `env.runCustomEffect`
